fix : 그룹 기능 수정 [노은빈]

+++ backend/controllers/group.js
@@ -3,13 +3,24 @@ const User = require('../models/user');
 
 // group 생성
 exports.create = async (req, res, next) => {
-  console.log('group',req.body);
   try {
-    await Group.create({
+    const existGroup = await Group.findOne({
+      where: {
+        name: req.body.name,
+        userID: req.user.id,
+      },
+    });
+
+    if (existGroup) {
+      return res.status(400).json({ message: "이미 존재하는 그룹 이름입니다" });
+    }
+
+    const groups = await Group.create({
       name: req.body.name,
       userID: req.user.id
     });
-    res.send(req.body);
+
+    res.send(groups);
   } catch (error) {
     console.error(error);
     return next(error);
@@ -21,13 +32,13 @@ exports.create = async (req, res, next) => {
 exports.index = async (req, res, next) => {
   const user = req.user;
   try {
-    const groups = await Group.findAll({
+    const existGroup = await Group.findAll({
       where : {
         userID : user.id,
       }
     });
-    console.log(groups);
-    res.send(groups);
+
+    res.status(200).send(existGroup);
   } catch (error) {
     console.error(error);
     return next(error);
@@ -38,14 +49,22 @@ exports.index = async (req, res, next) => {
 // group 이름 조회
 exports.read = async (req, res, next) => {
   try {
-    const groups = await Group.findOne({
+    const existGroup = await Group.findOne({
       where: { 
         id: req.params.id,
         userID : req.user.id
       }
     });
-    console.log(groups);
-    res.send(groups);
+
+    if (!existGroup) {
+      return res.status(404).json({ message: "group 정보를 찾을 수 없습니다" });
+    }
+
+    if (existGroup.userID !== req.user.id) {
+      return res.status(403).json({ message: "다른 사용자의 정보입니다" });
+    }
+
+    res.status(200).send(existGroup);
   } catch (error) {
     console.error(error);
     return next(error);
@@ -56,15 +75,41 @@ exports.read = async (req, res, next) => {
 // group 수정
 exports.update = async (req, res, next) => {
   try {
-    await Group.update({
+    const existGroup = await Group.findOne({
+      where: {
+        id: req.params.id,
+      },
+    });
+
+    if (!existGroup) {
+      return res.status(404).json({ message: "해당 Group이 없습니다" });
+    }
+
+    if (existGroup.userID !== req.user.id) {
+      return res.status(403).json({ message: "다른 사용자의 정보입니다" });
+    }
+
+    const confirmGroup = await Group.findOne({
+      where: {
+        name: req.body.name,
+        userID: req.user.id,
+      },
+    });
+
+    if (confirmGroup && confirmGroup.id !== existGroup.id) {
+      return res.status(400).json({ message: "이미 존재하는 그룹 이름입니다" });
+    }
+
+    const group = await Group.update({
       name: req.body.name,
     },{
       where: { 
         id: req.params.id,
         userID : req.user.id
       }
-      // where: { id: req.params.id }, // url에 포함된 정보를 가져올 때는 params에서 가져옴
     });
+
+    res.status(200).json({ message: "그룹이 수정되었습니다" });
   } catch (error) {
     console.error(error);
     return next(error);
@@ -72,10 +117,23 @@ exports.update = async (req, res, next) => {
   res.end();
 };
 
-
 // group 삭제
 exports.remove = async (req, res, next) => {
   try {
+    const existGroup = await Group.findOne({
+      where: {
+        id: req.params.id,
+      },
+    });
+
+    if (!existGroup) {
+      return res.status(404).json({ message: "해당 group가 없습니다" });
+    }
+
+    if (existGroup.userID !== req.user.id) {
+      return res.status(403).json({ message: "로그인한 사용자가 아닙니다" });
+    }
+
     await Group.destroy({ 
       where: { 
         id: req.params.id,
@@ -83,6 +141,8 @@ exports.remove = async (req, res, next) => {
       },
       truncate: false,
     });
+    
+    res.status(200).json({ message: "group 정보가 삭제되었습니다." });
   } catch (error) {
     console.error(error);
     return next(error);
